src/requirements.txt
clickhouse-connect==0.8.17
PyJWT==2.10.1
mcp==1.6.0
